Renamed tags -> categorizationAttributes
Customer rationale: tags are typically synonymous with labels, that are to some degree "markers" for a feature. We have string enumerations that we can connect to capabilities, and might in the future have even more complex structures. Tags is too simple an analogy. This is used to categorize the entities in question in a manner orthogonal to the subsumption hierarchy; thus "categorization attributes".