Broaden domain of hasBuildingComponent
As indicated by our colleagues at SWECO, a generic relationship linking spaces and building components was missing (we've only had Building -> BuildingComponent, via hasBuildingComponent).

To resolve this, I've broadened the domain of the "hasBuildingProperty" to be Space rather than Building. This relaxes the semantics but does not make any old graphs inconsistent, so is a non-breaking change.

In addition, I've added the "componentOf" relationship going to opposite direction (BuildingComponent.componentOf -> Space), in case such traversal is needed. The pre-existing "componentOfBuilding" property is asserted as subproperty of this new propery; again, maintaining backwards compatibility.